Reformat with black

+++ pymanip/video/pco/__init__.py
@@ -8,8 +8,10 @@ import numpy as np
 from .camera import PCO_Camera
 from . import pixelfly as pf
 
+
 def PCO_read_binary_file(f):
-    return np.fromfile(f, dtype='<i2').reshape( (1200, 1600) )
+    return np.fromfile(f, dtype="<i2").reshape((1200, 1600))
+
 
 def print_available_pco_cameras():
     cams = list()
